+++ extras-examples/src/main/java/com/metsci/glimpse/extras/examples/charts/bathy/BathymetryExample.java
@@ -252,23 +252,23 @@ public class BathymetryExample
         dotPainter.setPointColor( 1, GlimpseColor.getBlack( ) );
 
         Vector2d pos = projection.project( LatLonGeo.fromDeg( 19.14, -80.23 ) );
-        Annotation a = annotationPainter.addAnnotation( "buoy 125A-3", ( float ) pos.getX( ), ( float ) pos.getY( ), 5, 2, HorizontalPosition.Left, VerticalPosition.Center, GlimpseColor.getGreen( ) );
+        Annotation a = annotationPainter.addAnnotation( "buoy 125A-3", pos.getX( ), pos.getY( ), 5, 2, HorizontalPosition.Left, VerticalPosition.Center, GlimpseColor.getGreen( ) );
         a.setBackgroundColor( GlimpseColor.getGray( 0.3f ) );
         dotPainter.addPointGeo( 1, 1, 19.14, -80.23, 0 );
 
         pos = projection.project( LatLonGeo.fromDeg( 18.88, -80.83 ) );
-        a = annotationPainter.addAnnotation( "buoy 126A-2", ( float ) pos.getX( ), ( float ) pos.getY( ), 5, 2, HorizontalPosition.Left, VerticalPosition.Center, GlimpseColor.getGreen( ) );
+        a = annotationPainter.addAnnotation( "buoy 126A-2", pos.getX( ), pos.getY( ), 5, 2, HorizontalPosition.Left, VerticalPosition.Center, GlimpseColor.getGreen( ) );
         a.setBackgroundColor( GlimpseColor.getGray( 0.3f ) );
         dotPainter.addPointGeo( 1, 1, 18.88, -80.83, 0 );
 
         pos = projection.project( LatLonGeo.fromDeg( 19.64, -79.50 ) );
-        a = annotationPainter.addAnnotation( "buoy 126A-1", ( float ) pos.getX( ), ( float ) pos.getY( ), 5, 2, HorizontalPosition.Left, VerticalPosition.Center, GlimpseColor.getRed( ) );
+        a = annotationPainter.addAnnotation( "buoy 126A-1", pos.getX( ), pos.getY( ), 5, 2, HorizontalPosition.Left, VerticalPosition.Center, GlimpseColor.getRed( ) );
         a.setBackgroundColor( GlimpseColor.getCyan( 0.3f ) );
         a.setBorderColor( GlimpseColor.getYellow( ) );
         dotPainter.addPointGeo( 1, 1, 19.64, -79.50, 0 );
 
         pos = projection.project( LatLonGeo.fromDeg( 19.80, -79.08 ) );
-        a = annotationPainter.addAnnotation( "buoy 125B-3", ( float ) pos.getX( ), ( float ) pos.getY( ), 5, 2, HorizontalPosition.Left, VerticalPosition.Center, GlimpseColor.getGreen( ) );
+        a = annotationPainter.addAnnotation( "buoy 125B-3", pos.getX( ), pos.getY( ), 5, 2, HorizontalPosition.Left, VerticalPosition.Center, GlimpseColor.getGreen( ) );
         a.setBorderColor( GlimpseColor.getYellow( ) );
         dotPainter.addPointGeo( 1, 1, 19.80, -79.08, 0 );
